ai_backend/requirements.txt
# Core Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
pydantic==2.10.4

# AI & LLM
langchain==0.3.20
langchain-google-genai==2.0.8
langgraph==0.2.62
langchain-community==0.3.15

# RAG
rank-bm25==0.2.2

# Web Search
langchain-tavily==0.1.6

# Utilities
websockets==14.1
python-multipart==0.0.20
